requirements.txt
Django==2.1.5
django-rest==0.0.1
djangorestframework==3.9.1
gunicorn==19.9.0
psycopg2-binary==2.7.7
pytz==2018.9
